move observing resolver to pkgtest
