+++ syft/pkg/cataloger/internal/pkgtest/observing_resolver.go
@@ -1,4 +1,4 @@
-package srctest
+package pkgtest
 
 import (
 	"fmt"
